refactor: separate rocksdb storage headers
