+++ persist/rocksdb/storage.h
@@ -1,0 +1,944 @@
+/**
+ * @file rocksdb_storage.h
+ * @brief RocksDB-backed storage implementation for the Bigtable emulator.
+ */
+
+#ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_BIGTABLE_EMULATOR_ROCKSDB_ROCKSDB_STORAGE_H
+#define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_BIGTABLE_EMULATOR_ROCKSDB_ROCKSDB_STORAGE_H
+
+#include "persist/storage.h"
+#include "persist/rocksdb/storage_row_tx.h"
+#include "persist/rocksdb/column_family_stream.h"
+#include "persist/rocksdb/filtered_table_stream.h"
+#include "filter.h"
+#include "persist/proto/storage.pb.h"
+#include "absl/strings/str_cat.h"
+#include "google/cloud/status.h"
+#include "google/cloud/status_or.h"
+#include "google/cloud/internal/make_status.h"
+#include <google/bigtable/admin/v2/table.pb.h>
+#include <google/bigtable/v2/data.pb.h>
+#include "rocksdb/db.h"
+#include "rocksdb/options.h"
+#include "rocksdb/slice.h"
+#include "rocksdb/utilities/transaction.h"
+#include "rocksdb/utilities/transaction_db.h"
+#include "rocksdb/iterator.h"
+#include "absl/types/optional.h"
+#include <algorithm>
+#include <chrono>
+#include <map>
+#include <set>
+#include <string>
+#include <string_view>
+#include <thread>
+#include <type_traits>
+#include <vector>
+
+namespace google {
+namespace cloud {
+namespace bigtable {
+namespace emulator {
+
+/**
+ * RocksDB-backed storage implementation for the Bigtable emulator.
+ */
+class RocksDBStorage : public Storage {
+  friend class RocksDBStorageRowTX;
+  friend class RocksDBColumnFamilyStream;
+ private:
+  /** Creates an iterator over the metadata column family (table keys). */
+  inline rocksdb::Iterator* createTablesIterator() const {
+    return db->NewIterator(roptions, metaHandle);
+  }
+ public:
+  /** Column family handle type. */
+  using CFHandle = rocksdb::ColumnFamilyHandle*;
+
+  /** Constructs storage from a RocksDB config. */
+  explicit RocksDBStorage(const storage::StorageRocksDBConfig& arg_storage_config) {
+    storage_config = arg_storage_config;
+  }
+
+  /** Constructs storage using FLAGS_storage_path and default meta CF name. */
+  explicit RocksDBStorage() {
+    storage_config = storage::StorageRocksDBConfig();
+    storage_config.set_db_path(absl::GetFlag(FLAGS_storage_path));
+    storage_config.set_meta_column_family("bte_metadata");
+  }
+
+  /** Destroys the storage and closes the database. */
+  virtual ~RocksDBStorage() {
+    Close();
+  }
+
+  /**
+   * Builds the RocksDB key for a (table, row, column_qualifier).
+   */
+  static inline std::string storageKey(
+      const std::string& table_name,
+      const std::string& row_key,
+      const std::string& column_qualifier
+  ) {
+    return absl::StrCat(table_name, "|", row_key, "|", column_qualifier);
+  }
+
+  /**
+   * Builds the RocksDB key prefix for a (table, row) to iterate all columns.
+   */
+  static inline std::string storageKeyPartial(
+      const std::string& table_name,
+      const std::string& row_key
+  ) {
+    return absl::StrCat(table_name, "|", row_key);
+  }
+
+  /**
+   * Parses a RocksDB key into (table_name, row_key, column_qualifier).
+   */
+  inline std::tuple<std::string, std::string, std::string> RawDataParseRowKey(
+      rocksdb::Iterator* iter
+  ) {
+    const auto key = iter->key().ToString();
+    std::vector<std::string_view> result;
+    auto left = key.begin();
+    for (auto it = left; it != key.end(); ++it)
+    {
+      if (*it == '|')
+      {
+        result.emplace_back(&*left, it - left);
+        left = it + 1;
+      }
+    }
+    if (left != key.end()) {
+      result.emplace_back(&*left, key.end() - left);
+    }
+    return std::make_tuple(std::string(result[0]), std::string(result[1]), std::string(result[2]));
+  }
+
+  /**
+   * Seeks the iterator to the first key for (table_name, row_prefix).
+   * RawData* methods implement lowest-level row storage; this seeks so we iterate a superset of the prefix.
+   */
+  inline void RawDataSeekPrefixed(
+      rocksdb::Iterator* iter,
+      const std::string& table_name,
+      const std::string& row_prefix
+  ) {
+    iter->Seek(rocksdb::Slice(storageKeyPartial(table_name, row_prefix)));
+  }
+
+  /**
+   * Puts serialized row data for (table, row, column) into RocksDB.
+   */
+  inline Status RawDataPut(
+      rocksdb::Transaction* txn,
+      const std::string& table_name,
+      const std::string& column_family,
+      const std::string& row_key,
+      const std::string& column_qualifier,
+      std::string&& data
+  ) {
+    const auto& cf = column_families_handles_map[column_family];
+    return GetStatus(txn->Put(cf, rocksdb::Slice(storageKey(table_name, row_key, column_qualifier)), rocksdb::Slice(std::move(data))), "Update commit row");
+  }
+
+  /**
+   * Gets serialized row data for (table, row, column) from RocksDB.
+   */
+  inline Status RawDataGet(
+      rocksdb::Transaction* txn,
+      const std::string& table_name,
+      const std::string& column_family,
+      const std::string& row_key,
+      const std::string& column_qualifier,
+      std::string* out
+  ) {
+    rocksdb::ReadOptions roptions;
+    const auto& cf = column_families_handles_map[column_family];
+    return GetStatus(txn->Get(roptions, cf, rocksdb::Slice(storageKey(table_name, row_key, column_qualifier)), out), "Load commit row", Status());
+  }
+
+  /**
+   * Deletes the underlying data for the given row's column.
+   */
+  inline Status RawDataDeleteColumn(
+      rocksdb::Transaction* txn,
+      const std::string& table_name,
+      const std::string& column_family,
+      const std::string& row_key,
+      const std::string& column_qualifier
+  ) {
+    const auto& cf = column_families_handles_map[column_family];
+    return GetStatus(txn->Delete(cf, rocksdb::Slice(storageKey(table_name, row_key, column_qualifier))), "Delete row column");
+  }
+
+  /**
+   * Deletes all columns for the given row in the column family (may be more expensive than RawDataDeleteColumn).
+   */
+  inline Status RawDataDelete(
+      rocksdb::Transaction* txn,
+      const std::string& table_name,
+      const std::string& column_family,
+      const std::string& row_key
+  ) {
+    const auto& cf = column_families_handles_map[column_family];
+    // Need to iterate through the columns
+    auto it = txn->GetIterator(roptions, cf);
+    for(it->Seek(rocksdb::Slice(storageKeyPartial(table_name, row_key))); it->Valid(); it->Next()) {
+      const auto& [k_table_name, k_row_key, _] = RawDataParseRowKey(it);
+      if (k_table_name != table_name || k_row_key != row_key) {
+        break;
+      }
+      auto status = GetStatus(txn->Delete(cf, it->key()), "Delete row");
+      if (!status.ok()) {
+        delete it;
+        return status;
+      }
+    }
+    delete it;
+    return Status();
+  }
+
+  /** @copydoc Storage::RowTransaction */
+  virtual std::unique_ptr<StorageRowTX> RowTransaction(std::string const& table_name, std::string const& row_key) {
+    return std::unique_ptr<RocksDBStorageRowTX>(new RocksDBStorageRowTX(table_name, row_key, StartRocksTransaction(), this));
+  }
+
+  /** @copydoc Storage::Close */
+  virtual Status Close() {
+    if (db == nullptr) {
+      return Status();
+    }
+
+    DBG("RocksDBStorage:Close killing meta handle");
+    for (auto& pair : column_families_handles_map) {
+      delete pair.second;
+    }
+    column_families_handles_map.clear();
+
+    rocksdb::WaitForCompactOptions opts;
+    opts.close_db = true;
+    DBG("RocksDBStorage:Close waiting for compact");
+    auto status = GetStatus(db->WaitForCompact(opts), "Wait for compact");
+    if (!status.ok()) {
+      DBG(absl::StrCat("RocksDBStorage:Close error: ", status.message()));
+      return status;
+    }
+    DBG("RocksDBStorage:Close deleting DB");
+    db = nullptr;
+    DBG("RocksDBStorage:Close deleted DB");
+    metaHandle = nullptr;
+    DBG("RocksDBStorage:Close exit");
+    return Status();
+  }
+
+  /** @copydoc Storage::Open */
+  virtual Status Open(std::vector<std::string> additional_cf_names = {}) {
+    options = rocksdb::Options();
+    txn_options = rocksdb::TransactionDBOptions();
+    woptions = rocksdb::WriteOptions();
+    roptions = rocksdb::ReadOptions();
+
+    options.create_if_missing = true;
+
+    std::vector<std::string> existing_cf_names;
+    auto status = rocksdb::TransactionDB::ListColumnFamilies(
+        options, storage_config.db_path(), &existing_cf_names);
+    bool is_new_database = !status.ok();
+    if (is_new_database) {
+      DBG("RocksDBStorage:Open no existing DB, will create new one");
+      existing_cf_names.clear();
+      // New database - start with just default CF
+      existing_cf_names.push_back(rocksdb::kDefaultColumnFamilyName);
+    }
+
+    // Build set of all column families we need
+    std::set<std::string> all_cf_set;
+    for (auto const& cf_name : existing_cf_names) {
+      all_cf_set.insert(cf_name);
+    }
+
+    // Add additional CFs from parameter
+    for (auto const& cf_name : additional_cf_names) {
+      all_cf_set.insert(cf_name);
+    }
+
+    // Always need meta CF
+    bool need_meta_cf = all_cf_set.find(storage_config.meta_column_family()) == all_cf_set.end();
+    if (need_meta_cf) {
+      all_cf_set.insert(storage_config.meta_column_family());
+    }
+
+    std::vector<rocksdb::ColumnFamilyDescriptor> column_families;
+    for (auto const& cf_name : all_cf_set) {
+      column_families.push_back(rocksdb::ColumnFamilyDescriptor(
+          cf_name, rocksdb::ColumnFamilyOptions()));
+    }
+
+    std::vector<CFHandle> handles;
+    auto open_status = OpenDBWithRetry(options, column_families, &handles);
+
+    // If we're creating a new DB and need meta CF, we need to do a two-step process
+    if (is_new_database && need_meta_cf && !open_status.ok()) {
+      std::vector<rocksdb::ColumnFamilyDescriptor> minimal_cfs;
+      minimal_cfs.push_back(rocksdb::ColumnFamilyDescriptor(
+          rocksdb::kDefaultColumnFamilyName, rocksdb::ColumnFamilyOptions()));
+
+      open_status = OpenDBWithRetry(options, minimal_cfs, &handles);
+      if (!open_status.ok()) {
+        DBG(absl::StrCat("RocksDBStorage:Open failed to create new DB: ", open_status.message()));
+        return open_status;
+      }
+
+      // Step 2: Add meta column family
+      CFHandle meta_cf_handle;
+      auto create_cf_status = db->CreateColumnFamily(
+          rocksdb::ColumnFamilyOptions(), storage_config.meta_column_family(), &meta_cf_handle);
+
+      if (!create_cf_status.ok()) {
+        DBG(absl::StrCat("RocksDBStorage:Open failed to create meta CF: ", create_cf_status.ToString()));
+        // Clean up
+        for (auto h : handles) delete h;
+        delete db;
+        db = nullptr;
+        return StorageError("Failed to create meta column family");
+      }
+
+      handles.push_back(meta_cf_handle);
+    } else if (!open_status.ok()) {
+      DBG(absl::StrCat("RocksDBStorage:Open failed: ", open_status.message()));
+      return open_status;
+    }
+
+    metaHandle = nullptr;
+    for (auto const& handle : handles) {
+      column_families_handles_map[handle->GetName()] = handle;
+      if (handle->GetName() == storage_config.meta_column_family()) {
+        metaHandle = handle;
+      }
+    }
+
+    if (metaHandle == nullptr) {
+      return StorageError("Meta column family not found after opening");
+    }
+
+    return Status();
+  }
+
+  /** @copydoc Storage::CreateTable */
+  virtual Status CreateTable(google::bigtable::admin::v2::Table& schema) {
+    auto txn = StartRocksTransaction();
+
+    auto key = rocksdb::Slice(schema.name());
+    if(KeyExists(txn, metaHandle, key)) {
+      return AlreadyExistsError("Table already exists.", GCP_ERROR_INFO().WithMetadata(
+          "table_name", schema.name()));
+    }
+
+    storage::TableMeta meta;
+    *meta.mutable_table() = schema;
+
+    std::vector<std::string> missing_cfs;
+    for(auto& cf : meta.table().column_families()) {
+      auto iter = column_families_handles_map.find(cf.first);
+      if (iter == column_families_handles_map.end()) {
+        missing_cfs.push_back(cf.first);
+      }
+    }
+
+    if (!missing_cfs.empty()) {
+      txn->Rollback();
+      delete txn;
+
+      // Close current database
+      Close();
+
+      // Open regular RocksDB (not TransactionDB) to add column families
+      rocksdb::DB* regular_db;
+      std::vector<rocksdb::ColumnFamilyDescriptor> existing_cfs;
+      std::vector<std::string> cf_names;
+      auto list_status = rocksdb::DB::ListColumnFamilies(options, storage_config.db_path(), &cf_names);
+      if (!list_status.ok()) {
+        return StorageError("Failed to list CFs during CF addition");
+      }
+
+      std::vector<CFHandle> cf_handles;
+      for (auto const& name : cf_names) {
+        existing_cfs.push_back(rocksdb::ColumnFamilyDescriptor(name, rocksdb::ColumnFamilyOptions()));
+      }
+
+      auto db_status = rocksdb::DB::Open(options, storage_config.db_path(), existing_cfs, &cf_handles, &regular_db);
+      if (!db_status.ok()) {
+        return StorageError("Failed to open regular DB for CF addition: " + db_status.ToString());
+      }
+
+      for (auto const& cf_name : missing_cfs) {
+        CFHandle new_handle;
+        auto create_status = regular_db->CreateColumnFamily(rocksdb::ColumnFamilyOptions(), cf_name, &new_handle);
+        if (!create_status.ok()) {
+          for (auto h : cf_handles) delete h;
+          delete regular_db;
+          return StorageError("Failed to create column family " + cf_name + ": " + create_status.ToString());
+        }
+        cf_handles.push_back(new_handle);
+      }
+
+      for (auto h : cf_handles) delete h;
+      delete regular_db;
+
+      auto reopen_status = Open();
+      if (!reopen_status.ok()) {
+        return reopen_status;
+      }
+
+      txn = StartRocksTransaction();
+
+      missing_cfs.clear();
+      for(auto& cf : meta.table().column_families()) {
+        auto iter = column_families_handles_map.find(cf.first);
+        if (iter == column_families_handles_map.end()) {
+          missing_cfs.push_back(cf.first);
+        }
+      }
+    }
+
+    auto status = GetStatus(txn->Put(metaHandle, key, rocksdb::Slice(SerializeTableMeta(meta))), "Put table key to metadata cf");
+
+    status = Commit(txn);
+    if (!status.ok()) {
+      return status;
+    }
+
+    // User needs to restart emulator with pre-created column families
+    if (!missing_cfs.empty()) {
+      std::string missing_list;
+      for (auto const& cf : missing_cfs) {
+        if (!missing_list.empty()) missing_list += ", ";
+        missing_list += cf;
+      }
+      return InvalidArgumentError(
+          "Column families do not exist. Please restart the emulator to create them.",
+          GCP_ERROR_INFO()
+              .WithMetadata("missing_column_families", missing_list)
+              .WithMetadata("table_name", schema.name()));
+    }
+
+    // TODO: Implement garbage collection for old cell versions based on gc_rule
+    return Status();
+  }
+
+  /** @copydoc Storage::DeleteTable */
+  virtual Status DeleteTable(std::string table_name, std::function<Status(std::string, storage::TableMeta)>&& precondition_fn) const {
+    auto txn = StartRocksTransaction();
+
+    std::string out;
+    auto status = GetStatus(
+        txn->Get(roptions, metaHandle, rocksdb::Slice(table_name), &out),
+        "Get table",
+        NotFoundError("No such table.", GCP_ERROR_INFO().WithMetadata(
+            "table_name", table_name)));
+    if (!status.ok()) {
+      return Rollback(txn, status);
+    }
+    const auto meta = DeserializeTableMeta(std::move(out));
+    if (!meta.ok()) {
+      return Rollback(txn, meta.status());
+    }
+    const auto precondition_status = precondition_fn(table_name, meta.value());
+    if (!precondition_status.ok()) {
+      return Rollback(txn, precondition_status);
+    }
+
+    const auto delete_status = GetStatus(
+        txn->Delete(metaHandle, rocksdb::Slice(table_name)),
+        "Delete table",
+        NotFoundError("No such table.", GCP_ERROR_INFO().WithMetadata(
+            "table_name", table_name)));
+
+    if (!delete_status.ok()) {
+      return Rollback(txn, delete_status);
+    }
+    return Commit(txn);
+  }
+
+  /** @copydoc Storage::HasTable */
+  virtual bool HasTable(std::string table_name) const {
+    std::string _;
+    return db->KeyMayExist(roptions, metaHandle, rocksdb::Slice(table_name), &_) && db->Get(roptions, metaHandle, rocksdb::Slice(table_name), &_).ok();
+  }
+
+  /** @copydoc Storage::GetTable */
+  virtual StatusOr<storage::TableMeta> GetTable(std::string table_name) const {
+    std::string out;
+    auto status = GetStatus(
+        db->Get(roptions, metaHandle, rocksdb::Slice(table_name), &out),
+        "Get table",
+        NotFoundError("No such table.", GCP_ERROR_INFO().WithMetadata(
+            "table_name", table_name)));
+    if (!status.ok()) {
+      return status;
+    }
+    return DeserializeTableMeta(std::move(out));
+  }
+
+  /** @copydoc Storage::UpdateTableMetadata */
+  virtual Status UpdateTableMetadata(std::string table_name, storage::TableMeta const& meta) {
+    auto txn = StartRocksTransaction();
+    auto key = rocksdb::Slice(table_name);
+    auto serialized = SerializeTableMeta(meta);
+
+    auto status = GetStatus(
+        txn->Put(metaHandle, key, rocksdb::Slice(serialized)),
+        "Update table metadata");
+
+    if (!status.ok()) {
+      txn->Rollback();
+      delete txn;
+      return status;
+    }
+
+    return Commit(txn);
+  }
+
+  /** @copydoc Storage::EnsureColumnFamiliesExist */
+  virtual Status EnsureColumnFamiliesExist(std::vector<std::string> const& cf_names) {
+    std::vector<std::string> missing_cfs;
+    for (auto const& cf_name : cf_names) {
+      auto iter = column_families_handles_map.find(cf_name);
+      if (iter == column_families_handles_map.end()) {
+        missing_cfs.push_back(cf_name);
+      }
+    }
+
+    if (missing_cfs.empty()) {
+      return Status();
+    }
+
+    for (auto& pair : column_families_handles_map) {
+      delete pair.second;
+    }
+    column_families_handles_map.clear();
+    delete db;
+    db = nullptr;
+
+    rocksdb::DB* regular_db;
+    std::vector<rocksdb::ColumnFamilyDescriptor> existing_cfs;
+    std::vector<std::string> existing_cf_names;
+    auto list_status = rocksdb::DB::ListColumnFamilies(options, storage_config.db_path(), &existing_cf_names);
+    if (!list_status.ok()) {
+      return StorageError("Failed to list CFs during CF addition");
+    }
+
+    std::vector<CFHandle> cf_handles;
+    for (auto const& name : existing_cf_names) {
+      existing_cfs.push_back(rocksdb::ColumnFamilyDescriptor(name, rocksdb::ColumnFamilyOptions()));
+    }
+
+    auto db_status = rocksdb::DB::Open(options, storage_config.db_path(), existing_cfs, &cf_handles, &regular_db);
+    if (!db_status.ok()) {
+      return StorageError("Failed to open regular DB for CF addition: " + db_status.ToString());
+    }
+
+    for (auto const& cf_name : missing_cfs) {
+      CFHandle new_handle;
+      auto create_status = regular_db->CreateColumnFamily(rocksdb::ColumnFamilyOptions(), cf_name, &new_handle);
+      if (!create_status.ok()) {
+        for (auto h : cf_handles) delete h;
+        delete regular_db;
+        return StorageError("Failed to create column family " + cf_name + ": " + create_status.ToString());
+      }
+      cf_handles.push_back(new_handle);
+    }
+
+    for (auto h : cf_handles) delete h;
+    delete regular_db;
+
+    return Open();
+  }
+
+  template<bool with_prefix, typename T>
+  class TablesMetadataView;
+
+  /**
+   * Forward iterator over table metadata (table name, TableMeta) in RocksDB metadata CF.
+   * @tparam with_prefix If true, only keys with the given prefix are iterated.
+   */
+  template<bool with_prefix>
+  class TablesMetadataIterator {
+   public:
+    using iterator_category = std::forward_iterator_tag;
+    using value_type = std::tuple<std::string, storage::TableMeta>;
+    using difference_type = std::ptrdiff_t;
+    using pointer = value_type;
+    using reference = value_type;
+   private:
+    template<bool p_with_prefix, typename T> friend class TablesMetadataView;
+
+    /** Underlying RocksDB iterator. */
+    rocksdb::Iterator* iter;
+    /** Cached (key, TableMeta) for current position. */
+    absl::optional<value_type> loaded_table_data;
+    /** Error from last DeserializeTableMeta. */
+    Status error;
+
+    /** When with_prefix is true, the key prefix to filter by. */
+    std::conditional_t<with_prefix == true, std::string, std::monostate> prefix;
+
+    /** Returns true if the iterator is valid (and optionally matches prefix). */
+    inline bool isValid() const {
+      if constexpr (with_prefix) {
+        return iter != nullptr && error.ok() && iter->Valid() && iter->key().starts_with(prefix);
+      } else {
+        return iter != nullptr && error.ok() && iter->Valid();
+      }
+    }
+
+    /** Loads (key, TableMeta) from current iterator position. */
+    inline void loadTableData() {
+      if (!isValid()) {
+        return;
+      }
+      auto meta = DeserializeTableMeta(iter->value().ToString());
+      if (!meta.ok()) {
+        error = meta.status();
+        return;
+      }
+      loaded_table_data = std::make_pair(iter->key().ToString(), meta.value());
+    }
+
+    template<
+        bool B = with_prefix,
+        typename std::enable_if<B, int>::type = 0
+    >
+    explicit TablesMetadataIterator(
+        rocksdb::Iterator*&& rocks_iterator,
+        const std::string& search_prefix
+    ): iter(std::move(rocks_iterator)), loaded_table_data(absl::nullopt), prefix(search_prefix) {
+      iter->Seek(rocksdb::Slice(prefix));
+      loadTableData();
+    };
+
+    template<
+        bool B = !with_prefix,
+        typename std::enable_if<B, int>::type = 0
+    >
+    explicit TablesMetadataIterator(
+        rocksdb::Iterator*&& rocks_iterator
+    ): iter(std::move(rocks_iterator)), loaded_table_data(absl::nullopt) {
+      iter->SeekToFirst();
+      loadTableData();
+    };
+
+    explicit TablesMetadataIterator(std::monostate _): iter(nullptr) {}
+
+   public:
+    /** Destroys the iterator and deletes the RocksDB iterator. */
+    ~TablesMetadataIterator() {
+      if (iter != nullptr) {
+        delete iter;
+      }
+    }
+
+    reference operator*() const { return loaded_table_data.value(); }
+    pointer operator->() { return loaded_table_data.value(); }
+
+    /** Prefix increment. */
+    TablesMetadataIterator& operator++() {
+      if (!isValid()) {
+
+        return *this;
+      }
+      iter->Next();
+      loadTableData();
+      return *this;
+    }
+
+    /** Postfix increment. */
+    TablesMetadataIterator operator++(int) { TablesMetadataIterator tmp = *this; ++(*this); return tmp; }
+
+    friend bool operator== (const TablesMetadataIterator& a, const TablesMetadataIterator& b) { return a.iter == b.iter || (!a.isValid() && !b.isValid()); };
+    friend bool operator!= (const TablesMetadataIterator& a, const TablesMetadataIterator& b) { return a.iter != b.iter && (a.isValid() || b.isValid()); };
+
+    /** Returns error from last load. */
+    Status status() const {
+      return error;
+    }
+  };
+
+  /**
+   * Range/view over table metadata (optionally filtered by key prefix).
+   * @tparam with_prefix If true, only tables with the given key prefix are included.
+   * @tparam T Storage type (e.g. RocksDBStorage) that provides createTablesIterator().
+   */
+  template<bool with_prefix, typename T>
+  class TablesMetadataView {
+   private:
+    /** When with_prefix, the key prefix. */
+    const std::conditional_t<with_prefix, std::string, std::monostate> prefix;
+    /** Pointer to the storage instance. */
+    const T* storage;
+
+    template<
+        bool B = with_prefix,
+        typename std::enable_if<B, int>::type = 0
+    >
+    explicit TablesMetadataView(const T* storage_ptr, const std::string prefix): storage(storage_ptr), prefix(std::move(prefix)) {};
+
+    template<
+        bool B = !with_prefix,
+        typename std::enable_if<B, int>::type = 0
+    >
+    explicit TablesMetadataView(const T* storage_ptr): storage(storage_ptr) {};
+
+
+    friend T;
+   public:
+    /** @cond */
+    template<
+        bool B = with_prefix,
+        typename std::enable_if<B, int>::type = 0
+    >
+    inline TablesMetadataIterator<true> begin() const {
+      return TablesMetadataIterator<true>(std::move(storage->createTablesIterator()), prefix);
+    }
+
+    template<
+        bool B = !with_prefix,
+        typename std::enable_if<B, int>::type = 0
+    >
+    inline TablesMetadataIterator<false> begin() const {
+      return TablesMetadataIterator<false>(std::move(storage->createTablesIterator()));
+    }
+    /** @endcond */
+
+    /** Returns past-the-end iterator. */
+    inline TablesMetadataIterator<with_prefix> end() const {
+      return TablesMetadataIterator<with_prefix>(std::monostate{});
+    }
+
+    /** Returns vector of table names in this view. */
+    inline std::vector<std::string> names() const {
+      std::vector<std::string> tmp;
+      DBG("TablesMetadataView:names listing tables");
+      std::transform(begin(), end(), std::back_inserter(tmp), [](auto el) -> auto {
+        DBG(absl::StrCat("TablesMetadataView:names => ", std::get<0>(el)));
+        return std::get<0>(el);
+      });
+      DBG("TablesMetadataView:names end");
+      return tmp;
+    }
+  };
+
+  /** Returns a view of all table metadata. */
+  inline TablesMetadataView<false, RocksDBStorage> Tables() const {
+    return TablesMetadataView<false, RocksDBStorage>(this);
+  }
+
+  /** Returns a view of table metadata with keys starting with prefix. */
+  inline TablesMetadataView<true, RocksDBStorage> Tables(const std::string& prefix) const {
+    return TablesMetadataView<true, RocksDBStorage>(this, prefix);
+  }
+
+  /** @copydoc Storage::DeleteColumnFamily */
+  virtual Status DeleteColumnFamily(std::string const& cf_name) override {
+    auto it = column_families_handles_map.find(cf_name);
+    if (it == column_families_handles_map.end()) {
+      return Status();  // Already deleted or doesn't exist
+    }
+    auto* handle = it->second;
+
+    // Drop from RocksDB (records drop in MANIFEST)
+    auto status = db->DropColumnFamily(handle);
+    if (!status.ok()) {
+      return GetStatus(status, "DropColumnFamily");
+    }
+
+    // Destroy the handle to free memory
+    status = db->DestroyColumnFamilyHandle(handle);
+    if (!status.ok()) {
+      return GetStatus(status, "DestroyColumnFamilyHandle");
+    }
+
+    // Remove from internal map so Destructor doesn't double-free
+    column_families_handles_map.erase(it);
+
+    return Status();
+  }
+
+  /** Returns a cell stream over all rows in the table. */
+  CellStream StreamTable(
+      std::string const& table_name
+  ) {
+    auto all_rows_set = std::make_shared<StringRangeSet>(StringRangeSet::All());
+    return StreamTable(table_name, all_rows_set);
+  }
+
+  /**
+   * Returns a cell stream over the table for the given row set.
+   * @param table_name Table name.
+   * @param range_set Row keys to include.
+   * @param prefetch_all_columns If true, prefetch all columns per row.
+   */
+  CellStream StreamTable(
+      std::string const& table_name,
+      std::shared_ptr<StringRangeSet> range_set,
+      bool prefetch_all_columns = false
+  ) {
+    std::vector<std::unique_ptr<AbstractFamilyColumnStreamImpl>> iters;
+    auto table = GetTable(table_name);
+    auto const& m = table.value().table().column_families();
+    for (auto const& it : m) {
+      const auto x = column_families_handles_map[it.first];
+      if (x == nullptr) {
+        continue;
+      }
+      std::unique_ptr<AbstractFamilyColumnStreamImpl> c = std::make_unique<RocksDBColumnFamilyStream>(it.first, x, range_set, this, table_name, prefetch_all_columns);
+      iters.push_back(std::move(c));
+    }
+    return CellStream(std::make_unique<StorageFitleredTableStream>(std::move(iters)));
+  }
+
+ private:
+  /** RocksDB configuration. */
+  storage::StorageRocksDBConfig storage_config;
+
+  /** RocksDB options. */
+  rocksdb::Options options;
+  /** Transaction DB options. */
+  rocksdb::TransactionDBOptions txn_options;
+  /** Write options. */
+  rocksdb::WriteOptions woptions;
+  /** Read options. */
+  rocksdb::ReadOptions roptions;
+  /** Transaction DB instance. */
+  rocksdb::TransactionDB* db;
+  /** Handle for the metadata column family. */
+  CFHandle metaHandle = nullptr;
+  /** Map from column family name to handle. */
+  std::map<std::string, CFHandle> column_families_handles_map;
+
+  /** Starts a new RocksDB transaction. */
+  inline rocksdb::Transaction* StartRocksTransaction() const {
+    return db->BeginTransaction(woptions);
+  }
+
+  /** Deserializes RowData from string. */
+  inline StatusOr<storage::RowData> DeserializeRow(std::string&& data) const {
+    storage::RowData row;
+    if (!row.ParseFromString(data)) {
+      return StorageError("DeserializeRow()");
+    }
+    return row;
+  }
+
+  /** Serializes RowData to string. */
+  inline std::string SerializeRow(storage::RowData row) const {
+    std::string out;
+    row.SerializeToString(&out);
+    return out;
+  }
+
+  /** Deserializes TableMeta from string. */
+  static inline StatusOr<storage::TableMeta> DeserializeTableMeta(std::string&& data) {
+    storage::TableMeta meta;
+    if (!meta.ParseFromString(data)) {
+      return StorageError("DeserializeTableMeta()");
+    }
+    assert(meta.has_table());
+    return meta;
+  }
+
+  /** Serializes TableMeta to string. */
+  inline std::string SerializeTableMeta(storage::TableMeta meta) const {
+    std::string out;
+    meta.SerializeToString(&out);
+    return out;
+  }
+
+  /** Returns true if the key exists in the column family. */
+  inline bool KeyExists(
+      rocksdb::Transaction* txn,
+      rocksdb::ColumnFamilyHandle* column_family,
+      const rocksdb::Slice& key
+  ) const {
+    std::string _;
+    return txn->Get(roptions, column_family, key, &_).ok();
+  }
+
+  /** Rolls back the transaction and returns the given status. */
+  inline Status Rollback(
+      rocksdb::Transaction* txn,
+      Status status
+  ) const {
+    const auto txn_status = txn->Rollback();
+    assert(txn_status.ok());
+    delete txn;
+    if (!status.ok()) {
+      return status;
+    }
+    return Status();
+  }
+
+  /** Commits the transaction. */
+  inline Status Commit(
+      rocksdb::Transaction* txn
+  ) const {
+    const auto status = txn->Commit();
+    assert(status.ok());
+    delete txn;
+    return Status();
+  }
+
+  /** Opens the DB with retries on IO error. */
+  inline Status OpenDBWithRetry(
+      rocksdb::Options options,
+      std::vector<rocksdb::ColumnFamilyDescriptor> column_families,
+      std::vector<rocksdb::ColumnFamilyHandle*>* handles
+  ) {
+    auto status = rocksdb::Status();
+    for(auto i = 0; i<5; ++i) {
+      status = rocksdb::TransactionDB::Open(options, txn_options, storage_config.db_path(), column_families, handles, &db);
+      if (status.IsIOError()) {
+        std::this_thread::sleep_for(std::chrono::milliseconds(250));
+        continue;
+      } else {
+        break;
+      }
+    }
+    return GetStatus(status, "Open database");
+  }
+
+  /** Builds a storage error status. */
+  static inline Status StorageError(std::string&& operation) {
+    return InternalError("Storage error",
+        GCP_ERROR_INFO().WithMetadata(
+            "operation", operation));
+  }
+
+  /** Converts RocksDB status; uses not_found_status when status is NotFound. */
+  inline Status GetStatus(rocksdb::Status status, std::string&& operation, Status&& not_found_status) const {
+    if (status.IsNotFound()) {
+      return not_found_status;
+    }
+    if (!status.ok()) {
+      return StorageError(std::move(operation));
+    }
+    return Status();
+  }
+
+  /** Converts RocksDB status to Storage Status. */
+  inline Status GetStatus(rocksdb::Status status, std::string&& operation) const {
+    if (!status.ok()) {
+      return StorageError(std::move(operation));
+    }
+    return Status();
+  }
+};
+
+}  // namespace emulator
+}  // namespace bigtable
+}  // namespace cloud
+}  // namespace google
+
+#endif  // GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_BIGTABLE_EMULATOR_ROCKSDB_ROCKSDB_STORAGE_H
